pulse/DataSync/S3Bucket/requirements.txt
# AWS SDK for Python
boto3==1.34.0
botocore==1.34.0

# FastAPI and dependencies
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.6.0

# Environment variables
python-dotenv==1.0.0
